requirements.txt
pandas==1.5.3
numpy==1.23.5
azureml-opendatasets==1.56.0
